fixes bug 43411 : having scale as a float in AWTGVTFont couldn't deliver enough bits into graphics2D.scale() at AWTGVTGlyphVector:926
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@731265 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/gvt/font/AWTGVTFont.java
@@ -44,9 +44,9 @@ import org.apache.batik.gvt.text.ArabicTextHandler;
  */
 public class AWTGVTFont implements GVTFont {
 
-    protected Font  awtFont;
-    protected float size;
-    protected float scale;
+    protected Font   awtFont;
+    protected double size;
+    protected double scale;
 
     /**
      * Creates a new AWTGVTFont that wraps the given Font.
@@ -66,7 +66,7 @@ public class AWTGVTFont implements GVTFont {
      * @param font The font object to wrap.
      * @param scale The scale factor to apply to font...
      */
-    public AWTGVTFont(Font font, float scale) {
+    public AWTGVTFont(Font font, double scale) {
         this.size = font.getSize2D()*scale;
         this.awtFont = font.deriveFont(FONT_SIZE);
         this.scale = size/awtFont.getSize2D();
@@ -223,7 +223,7 @@ public class AWTGVTFont implements GVTFont {
                                          int limit,
                                          FontRenderContext frc) {
         return new GVTLineMetrics
-            (awtFont.getLineMetrics(chars, beginIndex, limit, frc), scale);
+            (awtFont.getLineMetrics(chars, beginIndex, limit, frc), (float)scale);
     }
 
     /**
@@ -234,7 +234,7 @@ public class AWTGVTFont implements GVTFont {
                                          int limit,
                                          FontRenderContext frc) {
         return new GVTLineMetrics
-            (awtFont.getLineMetrics(ci, beginIndex, limit, frc), scale);
+            (awtFont.getLineMetrics(ci, beginIndex, limit, frc), (float)scale);
     }
 
     /**
@@ -242,7 +242,7 @@ public class AWTGVTFont implements GVTFont {
      *  FontRenderContext.
      */
     public GVTLineMetrics getLineMetrics(String str, FontRenderContext frc) {
-        return new GVTLineMetrics(awtFont.getLineMetrics(str, frc), scale);
+        return new GVTLineMetrics(awtFont.getLineMetrics(str, frc), (float)scale);
     }
 
     /**
@@ -253,33 +253,33 @@ public class AWTGVTFont implements GVTFont {
                                          int limit,
                                          FontRenderContext frc) {
         return new GVTLineMetrics
-            (awtFont.getLineMetrics(str, beginIndex, limit, frc), scale);
+            (awtFont.getLineMetrics(str, beginIndex, limit, frc), (float)scale);
     }
 
     /**
      * Returns the size of this font.
      */
     public float getSize() {
-        return size;
+        return (float)size;
     }
 
     /**
      * Returns the horizontal kerning value for this glyph pair.
      */
     public float getHKern(int glyphCode1, int glyphCode2) {
-        return 0f;
+        return 0.0f;
     }
 
     /**
      * Returns the vertical kerning value for this glyph pair.
      */
     public float getVKern(int glyphCode1, int glyphCode2) {
-        return 0f;
+        return 0.0f;
     }
 
     /////////////////////////////////////////////////////////////////////////
 
-    public static final float FONT_SIZE = 48f;
+    public static final float FONT_SIZE = 48.0f;
 
     /**
      * Returns the geometry of the specified character. This method also put
